"Mike's version"
Added separate controls for enabling marginal histograms on each axis.
Added Seperate bin count countrols for each axis.
Added Separate bin height scaling controls for each axis.

Reorganized control panel so axis-specific controls are at the top,
plot-specific (but axis generic) controls are next,
and program-wide (global generic) stuff is at the bottom.

Cleaned up code for creating and handling per axis controls, though there
is still a lot more to do.

Next on the agenda: equi-depth histograms, 1D density estimates from
histograms, and 2D joint and conditional densities.  Also, separating
out "normalization" from "transformation" For example, taking the log
of a variable, or taking its rank, is a "1D transformation", whereas
moving the view so the mean is in the middle and the +/- 3 sigma is on
screen is a "normalization".