Now plotting symbols (sprites) are specified with seprate menus for selected and unselected points (per plot).
This is yet one more step on the way to Real Multiple Brushes.

Also, I changed write_asci_file() to use the default float format.
This prints (small) integers as *integers*.  E.g., it prints the
number three as "3" instead of as "0.3000000e01" but it also prints
floats using 8 digits of precision, and goes to scientific notation
when necessary, including (I think) when the field width does not
allow full precision without scientific notation.  I.e. it should print
the number 0.00000012345678 as "1.2345678e-7" and not as "0.00000012"